restructure code, add refqc subcmd
